[Update] timeLimit Double로 변경

+++ src/main/java/com/example/codeArena/Problem/dto/ProblemCreateDto.java
@@ -15,7 +15,7 @@ public class ProblemCreateDto {
     private String outputFormat;
     private String sampleInput;
     private String sampleOutput;
-    private Integer timeLimit;
+    private Double timeLimit;
     private Integer memoryLimit;
     private String category;
     private String[] tags;
